non-root: enable non-root checkpoint/restore
This commit enables checkpointing and restoring of applications as
non-root.

First goal was to enable checkpoint and restore of the env00 test case.

This uses the information from opts.uid and opts.cap_eff to skip certain
code paths which do not work as non-root.

The whole kerndat cache mechanism is still skipped for now. All kerndat
tests working as non-root are being run, but no cache file is written as
the default kerndat location requires root to write it.

This might lead to longer startup times for CRIU as non-root, but as a
first step towards non-root checkpoint/restore it is skipped for now.

Signed-off-by: Adrian Reber <areber@redhat.com>