test: skip vt test if device does not exist
On Jenkins on s390x we are seeing following error:

./vt --pidfile=vt.pid --outfile=vt.out --filename=vt.test
make: *** [Makefile:432: vt.pid] Error 1
 Test zdtm/static/vt FAIL at ['make', '--no-print-directory', '-C', 'zdtm/static', 'vt.pid']
Test output: ================================
08:08:15.556:    54: ERR: vt.c:38: Open virtual terminal vt.test failed (errno = 6 (No such device or address))
08:08:15.556:    53: ERR: test.c:316: Test exited unexpectedly with code 1

Which is probably because the host does not have ttyS0.

This commit adds a checkskip file for the vt test to only run it if the
device used in the test can be opened on the current host.

Signed-off-by: Adrian Reber <areber@redhat.com>